修改bug   我的举报

+++ app/src/main/java/com/huaxixingfu/sqj/http/api/BarListApi.java
@@ -15,7 +15,7 @@ public final class BarListApi implements IRequestApi {
 
     public final static class Bean {
         public int sysBarType;
-        public int appGuideId;
+        public long appGuideId;
         public String sysBarTitle;
         public String appGuideImageUrl;
         public String sysBarImageUrl;
